Change StackSize to int from ushort, cast as ushort on network transmit

+++ Source/ACE.Server/Network/Structure/HousePayment.cs
@@ -49,7 +49,7 @@ namespace ACE.Server.Network.Structure
             WeenieID = wo.WeenieClassId;
             Name = wo.Name;
             PluralName = wo.NamePlural;
-            Num = wo.StackSize ?? 1;
+            Num = (uint)(wo.StackSize ?? 1);
         }
     }
